% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOBilateralTrade.R
\name{calcFAOBilateralTrade}
\alias{calcFAOBilateralTrade}
\title{calcFAOBilateralTrade}
\usage{
calcFAOBilateralTrade(
  output = "value",
  products = "kcr",
  prodAgg = TRUE,
  harmonize = FALSE,
  fiveYear = TRUE
)
}
\arguments{
\item{output}{"value", "qty", or "price"}

\item{products}{"kcr", "kli", or "kothers"}

\item{prodAgg}{binary to keep FAO product level or magpie}

\item{harmonize}{combine input and export sheets with harmonization algorithm
 based on reliability index(Gelhar 1996)
default off as there are some big differences
and perhaps using the import sheet is better overall i.e. more matching with massbal}

\item{fiveYear}{if TRUE calculate only 5 years (1995 to 2020, 5-year steps) to reduce memory load}
}
\value{
List of magpie objects with results on bilateral country level,
weight on bilateral country level, unit and description.
}
\description{
Calculates bilateral trade values based on FAO trade matrix
}
\examples{
\dontrun{
calcOutput("FAOBilateralTrade", output = "qty", products = "kcr")
}

}
\author{
David M Chen
}
