% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOharmonized.R
\name{calcFAOharmonized}
\alias{calcFAOharmonized}
\title{calcFAOharmonized}
\usage{
calcFAOharmonized(src = "pre2010", output = "FB")
}
\arguments{
\item{src}{source "pre2010" or "post2010" "pre2010" returns the FAO sheet that runs
until 2013 (actually 2013, but gets effectively chopped to 2010 by massbalance). "post2010" combines
the new FAO FB, SUA, and CB sheets that were re-done for 2010 onwards.}

\item{output}{whether to return FB (Food balance sheet) or SUA (SUpply Utilization Accounts)}
}
\value{
FAO harmonized data, weight as NULL, and a description as as a list of MAgPIE objects
}
\description{
Calculate harmonized FAO Commodity Balance and Food Supply data based on CB, only harvested areas
             are taken from ProdSTAT. This functions adds the CBCrop, CBLive, FSCrop and FSLive data together.
}
\examples{
\dontrun{
a <- calcOutput("FAOharmonized")
}
}
\author{
Ulrich Kreidenweis, David Chen, Kristine Karstens
}
