% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFodder2010.R
\name{convertFodder2010}
\alias{convertFodder2010}
\title{Convert Fodder2010 data}
\usage{
convertFodder2010(x)
}
\arguments{
\item{x}{MAgPIE object containing original values}
}
\value{
Data as MAgPIE object with common country list
}
\description{
Converts new FAO Fodder2010 data to fit to the common country list.
Data starts in 2010 so no historical country transitions are needed,
but standard ones are included defensively. Units are kept as tonnes and ha.
}
\examples{
\dontrun{
a <- readSource("Fodder2010", convert = TRUE)
}
}
\seealso{
[readFodder2010()], [readSource()]
}
\author{
David Chen
}
