% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFAO_FRA2025.R
\name{convertFRA2025}
\alias{convertFRA2025}
\title{Convert FRA2025 data}
\usage{
convertFRA2025(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing original values}

\item{subtype}{The FAO FRA 2025 subtype.}
}
\value{
Data as MAgPIE object with common country list
}
\description{
Convert FRA2025 data
}
\examples{
\dontrun{
a <- readSource("FRA2025", "growing_stock", convert = TRUE)
}

}
\seealso{
[readSource()]
}
\author{
Simin Yu
}
