% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFAO_online.R
\name{downloadFAO_online}
\alias{downloadFAO_online}
\title{Download FAO data}
\usage{
downloadFAO_online(subtype)
}
\arguments{
\item{subtype}{Type of FAO data that should be read.}
}
\description{
Downloads the latest data and meta data form the FAOStat website.
}
\author{
Debbora Leip, David Chen, Bin Lin
}
