% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFAO_FRA2025.R
\name{downloadFRA2025}
\alias{downloadFRA2025}
\title{Download FAO FRA data}
\usage{
downloadFRA2025()
}
\description{
Downloads the FRA 2025 bulk data from the FAO-FRA website available at 
https://fra-data.fao.org/assessments/fra/2025/WO/data-download/
Removes month and date from the names of three main files
(FRA_Years_2025, Intervals_2025 and Annual_2025.csv).
}
