% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrfaocore-package.R
\docType{package}
\name{mrfaocore-package}
\alias{mrfaocore}
\alias{mrfaocore-package}
\title{mrfaocore: madrat-based package providing core FAO-related preprocessing functions}
\description{
This madrat-based package provides core FAO-related preprocessing functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrfaocore}
}

}
\author{
\strong{Maintainer}: David Chen \email{david.chen@pik-potsdam.de}

Authors:
\itemize{
  \item David Chen \email{david.chen@pik-potsdam.de}
  \item Ulrich Kreidenweis
  \item Abhijeet Mishra
  \item Kristine Karstens
  \item Benjamin Leon Bodirsky
  \item Debbora Leip
  \item Mishko Stevanovic
  \item Benjamin Leon Bodrisky
  \item David Klein
  \item Edna Molina Bacca
}

}
\keyword{internal}
