% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFodder2010.R
\name{readFodder2010}
\alias{readFodder2010}
\title{Read Fodder2010}
\usage{
readFodder2010()
}
\value{
FAO fodder data as MAgPIE object
}
\description{
Read in new FAO fodder data (2010-2023) downloaded from the FAO SWS system.
This dataset uses CPC item codes and M49 country codes.
Contains elements: Area Harvested (5312, ha) and Production (5510, tonnes).
}
\examples{
\dontrun{
a <- readSource("Fodder2010")
}
}
\seealso{
[readSource()]
}
\author{
David Chen
}
