% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAO_FRA2025.R
\name{readFRA2025}
\alias{readFRA2025}
\title{Read FRA2025}
\usage{
readFRA2025(subtype)
}
\arguments{
\item{subtype}{data subtype. Available subtypes: forest_area, deforestation, growing_stock, biomass_stock,
carbon_stock, management, disturbance, forest_fire, expansion, afforestation, net_change}
}
\value{
Magpie object of the FRA 2025 data
}
\description{
Read-in an FRA (forest resource assessment) dataset from 2025. 
Three subtypes are added compared to FRA2020: expansion, afforestation, net_change.
Three files are required for running: FRA_Years_2025.csv, Annual_2025.csv and Intervals_2025.csv.
}
\examples{
\dontrun{
a <- readSource("FRA2025", "growing_stock")
}

}
\seealso{
[readSource()]
}
\author{
Simin Yu
}
