% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrfaocore-package.R
\docType{package}
\name{mrfaocore-package}
\alias{mrfaocore}
\alias{mrfaocore-package}
\title{mrfaocore: One-line description of this awesome package}
\description{
One-paragraph description of this awesome package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrfaocore}
}

}
\author{
\strong{Maintainer}: Pascal Sauer \email{pascal.sauer@pik-potsdam.de}

}
\keyword{internal}
