% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAO_online.R
\name{readFAO_online}
\alias{readFAO_online}
\title{Read FAO_online}
\usage{
readFAO_online(subtype)
}
\arguments{
\item{subtype}{Type of FAO data that should be read. Available types are:
\itemize{
\item `CBCrop`: Commodity Balance Crop (CommodityBalances_Crops_E_All_Data.zip)
\item `CBLive`: Commoditiy Balance Livestock (CommodityBalances_LivestockFish_E_All_Data.zip)
\item `Crop`: Production Crops ("Production_Crops_E_All_Data.zip")
\item `CropProc`: Production Crops Processed ("Production_CropsProcessed_E_All_Data.zip")
\item `FB2010`: New Food Balance Sheet, 2010 onwards ("FoodBalanceSheets_E_All_Data.zip")
\item `SUA2010`: New Supply Utilization Accounts, 2010 onwards ("SUA_Crops_Livestock_E_All_Data_(Normalized).zip")
\item `CB2010`: New Commodity Balance Sheets, 2010 onwards ("CommodityBalances_(non-food)_(2010-)_E_All_Data_(Normalized).zip") #nolint
\item `Fertilizer`: Fertilizer ("Resources_Fertilizers_E_All_Data.zip")
\item `FertilizerProducts`: Fertilizer by product ("Inputs_FertilizersProduct_E_All_Data_(Normalized).zip")
\item `FertilizerNutrients`: Fertilizer by nutrient ("Inputs_FertilizersNutrient_E_All_Data_(Normalized).zip")
\item `Fodder`: Fodder (data that has been manually downloaded from the FAOSTAT website as
seperate .xls files via a search for "forage" and "fodder" withing
Production-Crops. These datasets have been added together to a "Fodder.csv" file)
\item `FoodSecurity`: Food Security Data ("Food_Security_Data_E_All_Data.zip")
\item `FSCrop`: Food Supply Crops ("FoodSupply_Crops_E_All_Data.zip")
\item `FSLive`: Food Supply Livestock ("FoodSupply_LivestockFish_E_All_Data.zip")
\item `Land`: Land ("Resources_Land_E_All_Data.zip")
\item `LiveHead`: Production Live Animals ("Production_Livestock_E_All_Data.zip")
\item `LivePrim`: Production Livestock Primary ("Production_LivestockPrimary_E_All_Data.zip")
\item `LiveProc`: Production Livestock Processed ("Production_LivestockProcessed_E_All_Data.zip")
\item `Pop`: Population ("Population_E_All_Data.zip")
\item `ForestProdTrade`: Forestry Production and Trade ("Forestry_E_All_Data_(Normalized).zip")
\item `PricesProducerAnnual`: Producer Prices - Annual ("Prices_E_All_Data.zip")
\item `PricesProducerAnnualLCU`: Producer Prices - Annual in LCU ("Prices_E_All_Data.zip")
\item `ValueOfProd`: Value of Agricultural Production ("Value_of_Production_E_All_Data.zip")
\item `ValueShares`: Value shares by industry and primary factors
\item `Trade`: Trade quantities and values
}}
}
\value{
FAO data as MAgPIE object
}
\description{
Read in FAO data that has been downloaded from the FAOSTAT website.
Files with exception of fodder.csv are aquired according to downloadFAO.
}
\details{
Update 23-Jan-2017 - Added FAO Forestry production and trade data (Abhi)
}
\examples{
\dontrun{
a <- readSource("FAO_online", "Crop")
}
}
\seealso{
[readSource()]
}
\author{
Ulrich Kreidenweis, Abhijeet Mishra, Mishko Stevanovic, David Klein, Edna Molina Bacca
}
